shell

+++ Kernel/src/sirqDispatcher.c
@@ -26,16 +26,16 @@ static uint64_t syscall_13 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
 static uint64_t syscall_14 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
 static uint64_t syscall_15 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
 static uint64_t syscall_16 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_17 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_18 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_19 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_20 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_21 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_22 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_23 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_24 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_25 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
-static int syscall_26 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_17 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_18 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_19 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_20 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_21 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_22 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_23 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_24 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_25 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
+static uint64_t syscall_26 (uint64_t rdi, uint64_t rsi, uint64_t rdx);
 
 extern void hang(); // Ubicada en loader.asm
 
@@ -100,7 +100,7 @@ uint64_t syscall_08 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 uint64_t syscall_09 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* si 'rdi' es cero: va a asignar 'rsi' bytes en memoria y devolver la direccion en la que los asignó.
 	   si no: va a liberar los bloques alocados en la direccion 'rsi' y devolver 0 si pudo hacerlo. */
-	return memory_handler( (char) rdi, rsi);
+	return (uint64_t) memory_handler( (char) rdi, rsi);
 }
 
 uint64_t syscall_10 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
@@ -136,7 +136,7 @@ uint64_t syscall_14 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* LIST PROCESSES 
 	*/
 
-	return list_processes_handler();
+	return (uint64_t) list_processes_handler();
 }
 
 uint64_t syscall_15 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
@@ -153,70 +153,70 @@ uint64_t syscall_16 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	return list_pipes_handler();
 }
 
-int syscall_17 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_17 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* Get calling PID
 	*/
 
 	return get_pid_handler();
 }
 
-int syscall_18 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_18 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* kill process with rdi PID.
 	*/
 
 	return kill_process_handler((int) rdi);
 }
 
-int syscall_19 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_19 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* Init Sem with rdi identifier, rsi initial count
 	*/
 
 	return init_sem_handler((char *) rdi, (int) rsi);
 }
 
-int syscall_20 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_20 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* open sem with rdi identifier
 	*/
 
 	return open_sem_handler((char *) rdi);
 }
 
-int syscall_21 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_21 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* close sem with rdi identifier.
 	*/
 
 	return close_sem_handler((char *) rdi);
 }
 
-int syscall_22 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_22 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* sem_wait with rdi identifier.
 	*/
 
 	return sem_wait_handler((char *) rdi);
 }
 
-int syscall_23 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_23 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* sem_post with rdi identifier.
 	*/
 
 	return sem_post_handler((char *) rdi);
 }
 
-int syscall_24 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_24 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* create pipe with rdi identifier.
 	*/
 
 	return create_pipe_handler((char *) rdi);
 }
 
-int syscall_25 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_25 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* write rsi content on pipe with rdi identifier.
 	*/
 
 	return write_pipe_handler((char *) rdi, (char *) rsi);
 }
 
-int syscall_26 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
+uint64_t syscall_26 (uint64_t rdi, uint64_t rsi, uint64_t rdx) {
 	/* read content on pipe with rdi identifier and store it on rsi string
 	*/
